Add extract_bspwm_canvas
The bspc command "bspc wm -l" really wants the JSON to be in a
particular order, which "bspc wm -d" happens output.  This means that we
can't use standard JSON tools manipulate the JSON because we can't
guarantee the order of the output.

Unfortunately, the bspwn extract_canvas utility does just that, so use
this one instead which uses regexen to keep the order.  It's somewhat
less robust, therefore, but seems to work well enough.