Track down slow input problems
The main problem is the blink tmux plugin.  It^s only really useful to
me if it can look at all the tmux panes, but this makes input lag far
too much for I've commented it out for now.

The other change is to remove completion on backspace which also slows
things down too much when making a lot of deletions.  Crtl-Space is
always available to start completions.